vscode-prerefactor: merge two files into downloads.ts
